Moves script files to separate folder
Moves all of the script files to the scripts folder and adjusts their references accordingly. This helps to clean up the repository and make finding relevant files easier.
